Fix payout type issues
